kernel-doc: Remove $decl_type='type name' hack
The $decl_type='type name' hack makes it impossible to document
macros with uppercase names (e.g. most of the macros in
object.h).

Now that we have explicitly tagged the struct and typedef doc
comments, we don't need that hack anymore.  This will make the
documentation for the  macros in object.h finally be rendered as
expected.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>