qdev: Don't register unreadable legacy properties
The existing code at qdev_class_add_legacy_property() will
register a property if both .print and .get are NULL, which is
useless as it will register a property that can't be read or
written.

The only PropertyInfo struct in the whole tree that has both
.print and .get set to NULL is qdev_prop_link.  This means every
link property in the code had a useless "legacy-" property being
registered.

Fix this by only registering legacy properties if a .print method
exists.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
---
This is a new patch in v3 of this series, it was previously
submitted as part of another series, at:
https://lore.kernel.org/qemu-devel/20201104172512.2381656-4-ehabkost@redhat.com
---
Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: "Daniel P. Berrangé" <berrange@redhat.com>
Cc: Eduardo Habkost <ehabkost@redhat.com>
Cc: qemu-devel@nongnu.org