qdev: Separate generic and device-specific property registration
qdev_class_add_property() and qdev_property_add_static() will
have code that's specific for device types.

object_class_property_add_field() and object_property_add_static()
will be generic and part of the QOM field property API.  Note
that the new functions have a `name` parameter because the plan
is to eventually get rid of the Property.name field.

The declarations for the new functions are being added to
qdev-properties.h, but they will be moved to a QOM header later.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
---
Changes v1 -> v2:
* Patch redone after changes in previous patches in the series
* Rename new functions to object*_property_add_field()
---
Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: "Daniel P. Berrangé" <berrange@redhat.com>
Cc: Eduardo Habkost <ehabkost@redhat.com>
Cc: qemu-devel@nongnu.org