qdev: Rename Property.name to Property.name_template
The Property.name field won't always be set, but it is very
easy to miss this detail when we see code using prop->name.

Rename the field to "name_template", to indicate it is just a
template used when creating new properties in some cases, but is
not always set.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
---
Changes v2 -> v3:
* Change name from .qdev_prop_name to .name_template.
  The property won't be qdev-specific, but it won't
  be always set.

This is a patch added in v2 of the series
---
Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: "Daniel P. Berrangé" <berrange@redhat.com>
Cc: Eduardo Habkost <ehabkost@redhat.com>
Cc: qemu-devel@nongnu.org