qdev: Don't set .name_template for array elements
All property code should be able to handle .name_template==NULL
without problems, so we don't need to set it for array elements.

This simple change will allow us to simplify the array property
registration code a lot in the next commits.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
---
This is a patch added in v2 of the series
---
Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: "Daniel P. Berrangé" <berrange@redhat.com>
Cc: Eduardo Habkost <ehabkost@redhat.com>
Cc: qemu-devel@nongnu.org