qdev: Make qdev_object_property_add() copy the Property struct
This will make the function more convenient to use, and allow
us to reuse it for array element properties.

I'm not adding a version of this function that doesn't copy the
struct, because this function isn't supposed to be used widely.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
---
This is a new patch added in v3 of the series.
---
Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: "Daniel P. Berrangé" <berrange@redhat.com>
Cc: Eduardo Habkost <ehabkost@redhat.com>
Cc: qemu-devel@nongnu.org