qom: FIELD_PROP macro
The FIELD_PROP macro can be used to define a Property struct when
initializing static variables, or as a function parameter (e.g.
as a paramter to object_class_property_add_field()).

The qdev-specific DEFINE_PROP macro can be defined using the new
macro.  The only difference is that DEFINE_PROP gets the property
name as the first argument.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
---
This is a new patch added in v3 of this series.
---
Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: "Daniel P. Berrangé" <berrange@redhat.com>
Cc: Eduardo Habkost <ehabkost@redhat.com>
Cc: qemu-devel@nongnu.org