smbios: entry-point-type option
Add command-line option that lets the SMBIOS entry point type to
be configured.

SMBIOS 3.0 support is necessary to allow us to support more
than 720 VCPUs in x86_64, due to SMBIOS 2.1 table size limits.

Note that it's still up to firmware to decide whether to generate
SMBIOS 2.1 and/or 3.0 entry points for the guest, using the
information contained in etc/smbios/smbios-anchor.  OVMF, for
example, is able to generate both entry points, depending on the
value of PcdSmbiosEntryPointProvideMethod.

The SMBIOS 3.0 entry point won't be enabled by default because it
is not supported yet by Seabios.  This may be changed once
Seabios starts supporting SMBIOS 3.0 entry points.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>