smbios: Support smbios-entry-point=auto
Support a new value for the smbios-entry-point option: "auto".
The new option can be used to keep compatibility with existing
VMs that use SMBIOS 2.1 entry points, but automatically using
SMBIOS 3.1 if the SMBIOS data is too large.

The new mode is not enabled by default because it would cause
confusion for users of Seabios (where SMBIOS 3.0 entry points are
not supported yet).  However, management software like libvirt
should be able to safely set smbios-entry-point=auto
automatically if using OVMF.