qdev: object_property_add_field() function
Extract the property registration code from
qdev_object_property_add(), so the generic property registration
code is kept separate from qdev-specific code.

This also makes Property.name_template optional when adding the
new property, allowing the function to be reused by the array
property registration code.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
---
This is a new patch added in v5 of the series.

This contains a portion of "qdev: Separate generic and
device-specific property registration".