test-char: Delete chardev correctly at char_file_test_internal()
commit 1e419ee68fa5 ("chardev: generate an internal id when none given")
changed the reference ownership semantics of
qemu_chardev_new(NULL, ...): now all chardevs created using
qemu_chardev_new() will be added to the /chardevs QOM container,
and the caller does not own a reference to the newly created
object.

However, the code at char_file_test_internal() had not been
updated and was calling object_unref() on a chardev object it
does not own.

Update the code to use object_unparent() at the end of the test,
to make sure the chardev object will be correctly deleted.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>