qom: Assert that objects being freed have no parent
QOM reference counting bugs are often hard to detect, but there's
one kind of bug that's easy to detect: if we are freeing an
object but is still attached to a parent, it means the reference
count is wrong (because the parent always hold a reference to all
its children).

Add an assert() to make sure we detect those cases.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>