cpu: Remove a CONFIG_USER_ONLY check on hw/core/cpu.h
There are function prototypes that were hidden from
CONFIG_USER_ONLY source files, but we don't need to do this if
the functions are simply not going to be available on
CONFIG_USER_ONLY binaries.

Removing the preprocessor conditional will allow us to add
CONFIG_USER_ONLY to poison.h in the future.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>